feat: 아이í경배아 히ì success clean build to mysql tables
