+++ src/main/java/com/scriptopia/demo/domain/EffectGradeDef.java
@@ -1,4 +1,4 @@
-package com.scriptopia.demo.dto;
+package com.scriptopia.demo.domain;
 
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -7,14 +7,16 @@ import lombok.Setter;
 @Entity
 @Getter
 @Setter
-public class ItemGradeDef {
+public class EffectGradeDef {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    private double weight;
+
+    private Long price;
+    private Double weight;
 
     @Enumerated(EnumType.STRING)
-    private Grade grade;  // enum 필드 추가
+    private Grade grade;
 }
