+++ src/main/java/com/scriptopia/demo/domain/GameSession.java
@@ -1,17 +1,14 @@
-package com.scriptopia.demo.dto;
+package com.scriptopia.demo.domain;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.time.LocalDateTime;
-
 
 @Entity
 @Getter
 @Setter
-public class LocalAccount {
-
+public class GameSession {
     @Id
     @GeneratedValue
     private Long id;
@@ -19,7 +16,7 @@ public class LocalAccount {
     @OneToOne(fetch = FetchType.LAZY)
     private User user;
 
-    private String email;
-    private String password;
-    private LocalDateTime updatedAt;
+    // 아직 mongoDB와 연동되는 지 모름 TYPE도 String(uuid)인지 LONG으로 할건지
+    private Long mongoId;
+
 }
