+++ src/main/java/com/scriptopia/demo/domain/LocalAccount.java
@@ -1,14 +1,17 @@
-package com.scriptopia.demo.dto;
+package com.scriptopia.demo.domain;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.time.LocalDateTime;
+
 
 @Entity
 @Getter
 @Setter
-public class GameSession {
+public class LocalAccount {
+
     @Id
     @GeneratedValue
     private Long id;
@@ -16,7 +19,7 @@ public class GameSession {
     @OneToOne(fetch = FetchType.LAZY)
     private User user;
 
-    // 아직 mongoDB와 연동되는 지 모름 TYPE도 String(uuid)인지 LONG으로 할건지
-    private Long mongoId;
-
+    private String email;
+    private String password;
+    private LocalDateTime updatedAt;
 }
