+++ src/main/java/com/scriptopia/demo/domain/MainStat.java
@@ -1,4 +1,4 @@
-package com.scriptopia.demo.dto;
+package com.scriptopia.demo.domain;
 
 public enum MainStat {
     STRENGTH,
