+++ src/main/java/com/scriptopia/demo/domain/PiaItem.java
@@ -1,5 +1,4 @@
-package com.scriptopia.demo.dto;
-
+package com.scriptopia.demo.domain;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -7,13 +6,15 @@ import jakarta.persistence.Id;
 import lombok.Getter;
 import lombok.Setter;
 
+
 @Entity
 @Getter
 @Setter
-public class TagDef {
-    @Id
-    @GeneratedValue
+public class PiaItem {
+    @Id @GeneratedValue
     private Long id;
 
-    private String tagName;
+    private String name;
+    private String price;
+    private String description;
 }
