+++ src/main/java/com/scriptopia/demo/domain/SocialAccount.java
@@ -1,10 +1,8 @@
-package com.scriptopia.demo.dto;
+package com.scriptopia.demo.domain;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
-
-import java.time.LocalDateTime;
 
 
 @Entity
