+++ src/main/java/com/scriptopia/demo/domain/TagDef.java
@@ -1,4 +1,5 @@
-package com.scriptopia.demo.dto;
+package com.scriptopia.demo.domain;
+
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -6,15 +7,13 @@ import jakarta.persistence.Id;
 import lombok.Getter;
 import lombok.Setter;
 
-
 @Entity
 @Getter
 @Setter
-public class PiaItem {
-    @Id @GeneratedValue
+public class TagDef {
+    @Id
+    @GeneratedValue
     private Long id;
 
-    private String name;
-    private String price;
-    private String description;
+    private String tagName;
 }
