+++ src/main/java/com/scriptopia/demo/domain/TradeStatus.java
@@ -1,4 +1,4 @@
-package com.scriptopia.demo.dto;
+package com.scriptopia.demo.domain;
 
 public enum TradeStatus {
     OWNED,
