+++ src/main/java/com/scriptopia/demo/domain/UserItem.java
@@ -1,22 +1,27 @@
-package com.scriptopia.demo.dto;
-
+package com.scriptopia.demo.domain;
 
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.Setter;
 
+
 @Entity
 @Getter
 @Setter
-public class UserCharacterImg {
+public class UserItem {
 
     @Id
     @GeneratedValue
     private Long id;
 
-
     @ManyToOne(fetch = FetchType.LAZY)
     private User user;
 
-    private String imgUrl;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private ItemDef itemDef;
+
+
+    private int remainingUses;
+    private TradeStatus tradeStatus;
 }
